refactor(service): dao, service 패키지 이동 및 인터페이스 구조 변경
